文件/requirements.txt
astroid==2.0.4
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
control==0.8.0
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.0.2
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
googletrans==2.4.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==19.9.0
gspread
httplib2==0.18.1
idna==2.10
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.16.0
matplotlib==3.0.0
mccabe==0.6.1
numpy==1.15.2
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==2.6.4
pandas==0.22.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pygsheets @ https://github.com/nithinmurali/pygsheets/archive/master.zip
pylint==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pydub
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.1.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.6.3
tornado==5.1.1
typed-ast==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.10.11
